ml/requirements-server.txt
tensorflow==2.15.0
flask==3.0.0
gunicorn==21.2.0
Pillow==10.1.0
numpy==1.26.2
